requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.1
black==23.9.1
bleach==6.0.0
blinker==1.6.2
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
contourpy==1.0.7
coverage==7.0.1
curl-cffi==0.5.5
cycler==0.11.0
Cython==0.29.33
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
depix @ git+https://github.com/beurtschipper/Depix@6dbbc29cd8e5ad7e52ddd18964e47b51ac2b54cb
deprecation==2.1.0
Django==4.1.7
django-crispy-forms==1.14.0
django-environ==0.9.0
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
executing==1.2.0
eyed3==0.9.7
fake-useragent==1.1.3
fastjsonschema==2.17.1
filetype==1.2.0
flake8==6.0.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
googletrans==3.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.16.3
httpx==0.23.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.6.0
install==1.3.5
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keyboard==0.13.5
kiwisolver==1.4.4
labelImg==1.8.6
libretranslatepy==2.1.1
line-profiler==4.0.2
lxml==4.9.2
magic-filter==1.0.9
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mimesis==11.1.0
mistune==3.0.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==7.0.1
multidict==6.0.4
mutagen==1.47.0
mypy==1.0.1
mypy-extensions==1.0.0
names==0.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.2
opencv-contrib-python==4.7.0.72
opencv-contrib-python-headless==4.7.0.72
opencv-python-headless==4.8.0.76
outcome==1.2.0
overrides==7.3.1
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.8.0
progressbar==2.5
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py-spy==0.3.14
pyaes==1.6.1
pyaimp==0.2.3
pyarrow==11.0.0
pyasn1==0.4.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.19.0
pydeck==0.8.1b0
pydub==0.25.1
pyflakes==3.0.1
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.2
Pympler==1.0.1
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.9
PyPasser==0.0.5
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
Pyrogram==2.0.97
pyrsistent==0.19.3
pyscreenshot==3.0
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
python-magic-bin==0.4.14
python-pptx==0.6.21
python-telegram-bot==20.1
python-youtube==0.9.0
pytweening==1.0.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
selenium==4.8.0
Send2Trash==1.8.2
simple-image-download==0.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundcloud-lib==0.6.1
soupsieve==2.4
SpeechRecognition==3.8.1
sqlparse==0.4.3
stack-data==0.6.2
streamlit==1.21.0
sympy==1.11.1
tbb==2019.0
telebot==0.0.5
telegram==0.0.1
Telethon==1.26.1
temp-email==0.2
temp-mail==0.2
termcolor==2.3.0
terminado==0.17.1
TgCrypto==1.2.5
tinycss2==1.2.1
tls-client==0.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.64.1
traitlets==5.9.0
translate==3.6.1
trio==0.22.0
trio-websocket==0.9.2
TwoCaptcha==0.0.1
types-Pillow==9.4.0.17
types-PyAutoGUI==0.9.3.5
types-PyScreeze==0.1.0.7
types-pywin32==305.0.0.9
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uri-template==1.3.0
urllib3==1.26.14
validators==0.20.0
vidgear==0.3.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
wikipedia==1.4.0
wikipedia-sections==2.0.0
wordlist==1.0.1
wsproto==1.2.0
XlsxWriter==3.0.9
yarl==1.8.2
youtube-dl==2021.12.17
yt-dlp==2023.10.13
zipp==3.15.0
